Move proposal downloader
